init and setup of scheduler working up

+++ database/migrations/2025_11_11_070212_create_appointments_table.php
@@ -10,40 +10,43 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('appointments', function (Blueprint $table) {
-            $table->id();
-            
-            // sheduler details
-            $table->foreignId('scheduler_id')->constrained()->onDelete('cascade');
-            $table->foreignId('property_id')->nullable()->constrained()->onDelete('set null');
-            
-            // date and time;
-            $table->date('date');
-            $table->time('start_time');
-            $table->time('end_time')->nullable();
-            
-            // location
-            $table->string('address');
-            $table->string('city');
-            $table->string('state');
-            $table->string('country');
-            $table->string('pin_code');
-            
-            //status of processing
-            $table->enum('status', ['pending', 'confirmed', 'cancelled', 'completed'])->default('pending');
-            
-            // assignment details
-            $table->foreignId('assigne_by')->nullable()->constrained('users');
-            $table->foreignId('assigne_to')->nullable()->constrained('users');
-            
-            //complete details
-            $table->foreignId('completed_by')->nullable()->constrained('users');
-            $table->foreignId('updated_by')->nullable()->constrained('users');
-            
-            // time-stamp
-            $table->foreignId('create_by')->nullable()->constrained('schedulers');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('appointments')) {
+            Schema::create('appointments', function (Blueprint $table) {
+                $table->id();
+
+                // sheduler details
+                $table->foreignId('scheduler_id')->constrained()->onDelete('cascade');
+                //comment out for now
+                // $table->foreignId('property_id')->nullable()->constrained()->onDelete('set null');
+
+                // date and time;
+                $table->date('date');
+                $table->time('start_time');
+                $table->time('end_time')->nullable();
+
+                // location
+                $table->string('address');
+                $table->string('city');
+                $table->string('state');
+                $table->string('country');
+                $table->string('pin_code');
+
+                //status of processing
+                $table->enum('status', ['pending', 'confirmed', 'cancelled', 'completed'])->default('pending');
+
+                // assignment details
+                $table->foreignId('assigne_by')->nullable()->constrained('users');
+                $table->foreignId('assigne_to')->nullable()->constrained('users');
+
+                //complete details
+                $table->foreignId('completed_by')->nullable()->constrained('users');
+                $table->foreignId('updated_by')->nullable()->constrained('users');
+
+                // time-stamp
+                $table->foreignId('create_by')->nullable()->constrained('schedulers');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
